Update ball's position every frame
When implementing, it became apparent that in order for the
FollowingAgent to have "smooth" behavior a new direction needed to be
added. This is a "none" direction indicating that the player should not
move at all.

Collision detection is not quite right for the player on the left-hand
side, but it should suffice for the present.