End game when a player loses
Direction was renamed to Action to more accurately reflect its meaning.
In order to simulate game-ending conditions, a new agent named
SedentaryAgent was added which does not move at all.

The collision detection needs to be fixed, but I'm on a plane and don't
feel like fetching my notebook to derive the proper condition to check
for.